requirements.txt
python-docx>=1.1.0
python-pptx>=1.0.0
openpyxl>=3.1.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
mypy>=1.0.0
